GUI: Implement new sorting algorithm
I had to remove incremental sorting/filtering which in theory causes a
decrease in performance because the sorting/filtering is not done in
the background anymore. However, I have not noticed any performance
problems.

I had to remove incremental sorting/filtering as the scroll position
would not go to the top after incremental sorting/filtering was
done. We somehow need to figure out whether incremental
sorting/filtering is done by attaching to some signal and then
updating the scrollbar position. Maybe do a
Notify on
https://docs.gtk.org/gtk4/method.FilterListModel.get_pending.html and
https://docs.gtk.org/gtk4/method.SortListModel.get_pending.html (pending
property for both) to check if it's 0. However that would also be
inefficient? Not sure what to do here. I think it's fine the way it is
done now anyways. Maybe let's not do premature optimizations.

Gtk states about incremental filtering/sorting:
`When your filter blocks the UI while filtering, you might consider turning this on. Depending on your model and filters, this may become interesting around 10,000 to 100,000 items.`