CMD: Allow negative serials
See: https://git.sr.ht/~jornane/php-openssl/commit/77161f97bf30f6429c31084375367aa372a40f4b

Previously the CA serial was randomly generated, giving a 50% chance
it was negative. This was fixed with the above commit. So this fix is
not ideal in the client here but we have to make sure it works with
older portals.